Some cleanup (#601)
Move filter and database update from MainActivity to application start. Migrate to assisted injection for media item and series screens

+++ core/src/androidMain/kotlin/com/minirogue/starwarscanontracker/core/model/room/dao/DaoMedia.kt
@@ -35,7 +35,7 @@ abstract class DaoMedia {
     abstract fun getMediaNotesById(mediaId: Long): Flow<MediaNotesDto>
 
     @Query("SELECT * FROM media_items WHERE id = :mediaID LIMIT 1")
-    abstract fun getMediaItemById(mediaID: Int): Flow<MediaItemDto>
+    abstract fun getMediaItemById(mediaID: Long): Flow<MediaItemDto>
 
     // The following are for MediaNotes interactions
     @Insert(onConflict = OnConflictStrategy.IGNORE)
